Delete todos and remove kernel pages.
