+++ flow-typed/npm/express_v4.16.x.js
@@ -160,7 +160,6 @@ declare class express$Route {
   checkout: express$RouteMethodType<this>;
   merge: express$RouteMethodType<this>;
 
-  // @TODO Missing 'm-search' but get flow illegal name error.
 
   notify: express$RouteMethodType<this>;
   subscribe: express$RouteMethodType<this>;
